moved section 1 to target nb, begin reworking section 2
